Add glassmorphism
But it doesn't work on firefox, so firefox user will see a transparent glass instead of blurred one, which is not bad, but not ideal either.
I tried implementing the -moz-element trick from this post https://medium.com/web-dev-survey-from-kyoto/glassmorphism-for-firefox-2113dff03747, but it's not working for unknown reason.
May come back to investigate further in the future